refactor: mv hmr logic to composable
